+++ packages/component-gateway/index.js
@@ -7,21 +7,22 @@ import { dumpDataUser } from '@qasir/auth/reducers';
 
 class ComponentGateway extends React.Component {
     componentDidMount() {
-        const {dispatch, currentPathName} = this.props
-        const auth = new Auth
-        const header = auth.headerLogin()
-        if (!header["Authorization"] || header["Authorization"] == "") {
-            dispatch(push("/login"))
+        const { dispatch, currentPathName } = this.props;
+        const auth = new Auth;
+        const header = auth.headerLogin();
+
+        if (!header['Authorization'] || header['Authorization'] == '') {
+            dispatch(push('/login'))
         } else {
             const hrefLocation = window.location.href
             const arr = hrefLocation.match(/\/#\/(.*)/g)
             const currentUrlPath = arr[0]
             dispatch(setShowMenuState(true))
             dispatch(dumpDataUser())
-            if (currentUrlPath != "/#/") {
-                dispatch(push(currentUrlPath.replace("/#", "")))
+            if (currentUrlPath != '/#/') {
+                dispatch(push(currentUrlPath.replace('/#', '')));
             } else {
-                dispatch(push("/dashboard"))
+                dispatch(push(''));
             }
         }
     }
@@ -37,9 +38,9 @@ class ComponentGateway extends React.Component {
 
 const mapStateToProps = ({routing, security}) => {
     return {
-        currentURL: routing.location != null ? routing.location.pathname : "/",
+        currentURL: routing.location != null ? routing.location.pathname : '/',
         clientSecretKey: 'mitra',
-        currentPathName: routing.location != null ? routing.location.pathname : "/"
+        currentPathName: routing.location != null ? routing.location.pathname : '/'
     }
 }