+++ packages/helpers/index.js
@@ -1,22 +1,22 @@
 class Helpers {
-    checkFormat(value) {
-        if (typeof(value)==='number') {
-            let val = (value / 1).toFixed(0).replace(',', '.');
-            return 'Rp ' + val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ".");
-        }
-        return value;
+  checkFormat(value) {
+      if (typeof(value)==='number') {
+          let val = (value / 1).toFixed(0).replace(',', '.');
+          return 'Rp ' + val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, '.');
+      }
+      return value;
 
-    }
-    formatelipsis(index) {
-        if (index.length > 40) {
-            return index.substr(0, 40) + '...';
-        }
-        return index;
+  }
+  formatelipsis(index) {
+      if (index.length > 40) {
+          return index.substr(0, 40) + '...';
+      }
+      return index;
 
-    }
-    replaceTagHTMLTotext (index){
-        return index.replace(/(<([^>]+)>)/ig, "");
-    }
+  }
+  replaceTagHTMLTotext (index){
+      return index.replace(/(<([^>]+)>)/ig, '');
+  }
 }
 
 export default Helpers;